message

+++ app/internal/domain/entity/link.go
@@ -3,6 +3,6 @@ package entity
 type Link struct {
 	ID           uint64 `json:"id"`
 	ChapterID    uint64 `json:"chapter_id"`
-	ParagraphNum uint64 `json:"paragraph_num"`
+	ParagraphNum uint32 `json:"paragraph_num"`
 	RID          uint64 `json:"r_id"`
 }
